moves the healthcare models and admin outside core project
