+++ healthcare/admin.py
@@ -1,7 +1,2 @@
 # Core admin
 from django.contrib import admin
-from .models import BloodPressure, BodyPhysique
-
-admin.site.register(BloodPressure)
-admin.site.register(BodyPhysique)
-
