+++ tracker/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from .models import BloodPressure, BodyPhysique
 
-# Register your models here.
+admin.site.register(BloodPressure)
+admin.site.register(BodyPhysique)
