+++ tracker/models.py
@@ -1,3 +1,112 @@
 from django.db import models
+from healthcare.models import HealthRecordCommonInfo
 
-# Create your models here.
+
+class BloodPressure(HealthRecordCommonInfo):
+    """
+        This is the model for a bloodpressure record
+    """
+
+    states = {
+        "normal": "Normal Blood Pressure",
+        "elevated": "Elevated Hypertension",
+        "high": "Hypertension Stage I",
+        "very_high": "Hypertension Stage II",
+        "risky": "Hypertension Crisis",
+        "low": "Hypotension"
+    }
+
+    systolic_pressure = models.IntegerField()
+    diastolic_pressure = models.IntegerField()
+
+    def __str__(self):
+        return f"{self.user}"
+
+    @property
+    def pressure(self):
+        """
+            A more readable format for blood pressure
+        """
+        return f"{self.systolic_pressure}/{self.diastolic_pressure}"
+
+    @property
+    def state(self):
+        # Based on
+        # https://www.heart.org/en/health-topics/high-blood-pressure/understanding-blood-pressure-readings
+
+        if ((self.systolic_pressure >= 90 and self.systolic_pressure < 120) and
+            (self.diastolic_pressure >= 60 and self.diastolic_pressure < 80)):
+            # Range is 90 - 119
+            # 60 - 79
+            return self.states.get('normal')
+        elif((self.systolic_pressure >= 120 and self.systolic_pressure < 130) and
+            (self.diastolic_pressure >= 60 and self.diastolic_pressure < 80)):
+            # Range is 120 - 129 and 60 - 79
+            return self.states.get('elevated')
+        elif(self.systolic_pressure > 180 or self.diastolic_pressure > 120):
+            # Calculate the risky here since this one should override things
+            # without constraints
+            return self.states.get('risky')
+        elif(self.systolic_pressure >= 140 or self.diastolic_pressure >= 90):
+            return self.states.get('very_high')
+        elif ((self.systolic_pressure >= 130 and self.systolic_pressure < 140) or
+              (self.diastolic_pressure >= 80 and self.diastolic_pressure < 90)):
+            # Range is 130 - 139 OR 80 - 89
+            return self.states.get('high')
+        elif (self.systolic_pressure < 90 or self.diastolic_pressure < 60):
+            # Range is 90 down and 60 down
+            return self.states.get('low')
+
+
+class BodyPhysique(HealthRecordCommonInfo):
+    """
+        This is the model for a the user's body physique record
+    """
+
+    states = {
+        "under": "Underweight",
+        "normal": "Normal weight",
+        "over": "Overweight",
+        "obese": "Obesity"
+    }
+
+    # Weight
+    weight_in_kilograms = models.IntegerField()
+    # Height
+    height_in_centimeters = models.IntegerField()
+
+    def __str__(self):
+        return f"{self.user}"
+
+    def _calculate_bmi(self):
+        """
+            Calculates bmi for interpratation
+            Based on:
+            https://www.cdc.gov/healthyweight/assessing/bmi/childrens_bmi/childrens_bmi_formula.html
+        """
+        return self.weight_in_kilograms / (self.height_in_meters**2)
+
+    @property
+    def height_in_meters(self):
+        """
+            Convert height height_in_centimeters to height_in_meters
+        """
+        return self.height_in_centimeters * 0.01
+
+    @property
+    def bmi_state(self):
+        """
+            Underweight = <18.5.
+            Normal weight = 18.5–24.9.
+            Overweight = 25–29.9.
+            Obesity = BMI of 30 or greater.
+        """
+        bmi = self._calculate_bmi()
+        if bmi < 18.5:
+            return self.states.get('under')
+        elif bmi >= 18.5 and bmi < 25:
+            return self.states.get('normal')
+        elif bmi >= 25 and bmi < 30:
+            return self.states.get('over')
+        elif bmi > 30:
+            return self.states.get('obese')
